[REFACTOR] Update divide components and style
컴포넌트와 스타일 분리
