--- client/src/components/common/Pagination.jsx
@@ -1,1 +1,0 @@
-import React from 'react';
